Tests accept supplementary groups from command line.
Tests now accept a list of supplementary groups as parameters to the
command line. For example, you can call test-bus with a list of group
parameters:

    sudo test-bus org.freedesktop.fwupd.update-hotplug-trusted hotplug wheel

Here the "hotplug" and "wheel" parameters are the supplementary groups
which are added to the process. Sudo is needed so that the groups can be
set (CAP_SETGID would be enough).

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>